client: add support for global configuration file for clients
Prior to this commit, users had to create a separate configuration file for
each process involved in  the coordinated checkpoint. This commit
enables the usage of a single config file for multiple processes. The
global config should be store inside `/etc/criu/criu-coordinator.json`
and it uses a slightly different format from the per-process config.

Example of global config:
```json
    {
       "address": "127.0.0.1",
       "port": "8080",
       "log-file": "/var/log/criu-coordinator.log",
       "dependencies": {
           "A": ["B", "C"],
           "B": ["C", "A"],
	   "C": ["A"]
       }
    }
```
Where `dependencies` field is a map of IDs (e.g: container IDs) to a list of dependencies.

Signed-off-by: Kouame Behouba Manasse <behouba@gmail.com>